respa-exchange: Move uploading functions to `uploader.py` for future symmetry
